requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
matplotlib==3.6.2
pandas==1.5.3
requests==2.28.1